+++ src/angular/chip/chip/chip.ts
@@ -1,0 +1,45 @@
+import { Directive, ElementRef, Input, NgZone, inject } from '@angular/core';
+import { booleanAttribute } from '@sbb-esta/lyne-angular/core';
+import type { SbbChipElement } from '@sbb-esta/lyne-elements/chip/chip.js';
+
+import '@sbb-esta/lyne-elements/chip/chip.js';
+
+@Directive({
+  selector: 'sbb-chip',
+})
+export class SbbChip {
+  #element: ElementRef<SbbChipElement> = inject(ElementRef<SbbChipElement>);
+  #ngZone: NgZone = inject(NgZone);
+
+  @Input()
+  public set value(value: string) {
+    this.#ngZone.runOutsideAngular(() => (this.#element.nativeElement.value = value));
+  }
+  public get value(): string {
+    return this.#element.nativeElement.value;
+  }
+
+  @Input({ transform: booleanAttribute })
+  public set readonly(value: boolean) {
+    this.#ngZone.runOutsideAngular(() => (this.#element.nativeElement.readonly = value));
+  }
+  public get readonly(): boolean {
+    return this.#element.nativeElement.readonly;
+  }
+
+  @Input({ transform: booleanAttribute })
+  public set negative(value: boolean) {
+    this.#ngZone.runOutsideAngular(() => (this.#element.nativeElement.negative = value));
+  }
+  public get negative(): boolean {
+    return this.#element.nativeElement.negative;
+  }
+
+  @Input({ transform: booleanAttribute })
+  public set disabled(value: boolean) {
+    this.#ngZone.runOutsideAngular(() => (this.#element.nativeElement.disabled = value));
+  }
+  public get disabled(): boolean {
+    return this.#element.nativeElement.disabled;
+  }
+}